Added BytesGeneratorFactory
